Define a protocol for GoalHealthKitConnection
This defines and documents a public interface for interacting with healthkit
metrics. That way internal methods defined by BaseGoalHealthKitConnection intended
for use by subclasses are no longer made available to all users.
